+++ webknossos/webknossos/dataset/layer.py
@@ -13,7 +13,6 @@ from cluster_tools import Executor
 from numpy.typing import DTypeLike
 from upath import UPath
 
-from ..client.context import _get_context
 from ..geometry import Mag, NDBoundingBox, Vec3Int, Vec3IntLike
 from ..geometry.mag import MagLike
 from ._array import ArrayException, TensorStoreArray
@@ -26,8 +25,8 @@ from ._downsampling_utils import (
     parse_interpolation_mode,
 )
 from ._upsampling_utils import upsample_cube_job
+from .attachments import Attachments
 from .data_format import DataFormat
-from .defaults import SSL_CONTEXT
 from .layer_categories import COLOR_CATEGORY, SEGMENTATION_CATEGORY, LayerCategoryType
 from .mag_view import MagView
 from .properties import (
@@ -46,6 +45,8 @@ if TYPE_CHECKING:
 
 from ..utils import (
     copytree,
+    dump_path,
+    enrich_path,
     get_executor_for_args,
     is_fs_path,
     is_remote_path,
@@ -200,45 +201,6 @@ def _find_mag_path(
         )
 
 
-def _enrich_mag_path(path: str, dataset_path: Path) -> Path:
-    upath = UPath(path)
-    if upath.protocol in ("http", "https"):
-        # To setup the mag for non-public remote paths, we need to get the token from the context
-        wk_context = _get_context()
-        token = wk_context.datastore_token
-        return UPath(
-            path,
-            headers={} if token is None else {"X-Auth-Token": token},
-            ssl=SSL_CONTEXT,
-        )
-
-    elif upath.protocol == "s3":
-        parsed_url = urlparse(str(upath))
-        endpoint_url = f"https://{parsed_url.netloc}"
-        bucket, key = parsed_url.path.split("/", maxsplit=1)
-
-        return UPath(
-            f"s3://{bucket}/{key}", client_kwargs={"endpoint_url": endpoint_url}
-        )
-
-    if not upath.is_absolute():
-        return resolve_if_fs_path(dataset_path / upath)
-    return resolve_if_fs_path(upath)
-
-
-def _dump_mag_path(path: Path, dataset_path: Path) -> str:
-    path = resolve_if_fs_path(path)
-    if str(path).startswith(str(dataset_path)):
-        return str(path).removeprefix(str(dataset_path)).lstrip("/")
-    if path.is_relative_to(dataset_path):
-        return str(path.relative_to(dataset_path))
-    if isinstance(path, UPath) and path.protocol == "s3":
-        return (
-            f"s3://{path.storage_options['client_kwargs']['endpoint_url']}/{path.path}"
-        )
-    return str(path)
-
-
 class Layer:
     def __init__(
         self, dataset: "Dataset", properties: LayerProperties, read_only: bool
@@ -291,7 +253,7 @@ class Layer:
             mag_path = (
                 _find_mag_path(resolved_path, mag.mag)
                 if mag.path is None
-                else _enrich_mag_path(mag.path, self.dataset.resolved_path)
+                else enrich_path(mag.path, self.dataset.resolved_path)
             )
             mag_is_read_only = read_only or _is_foreign_mag(
                 self.dataset.resolved_path, self.name, mag_path
@@ -397,20 +359,18 @@ class Layer:
 
         # The MagViews need to be updated
         for mag in self._mags.values():
-            if self.dataset.properties.version == 2:
-                if not mag.is_foreign:
-                    mag._properties.path = str(
-                        self.resolved_path.relative_to(self.dataset.resolved_path)
-                        / mag.path.name
-                    )
-                else:
-                    assert mag._properties.path is not None  # for type checking
-            else:
-                mag._path = (
-                    _enrich_mag_path(mag._properties.path, self.dataset.resolved_path)
-                    if mag._properties.path is not None
-                    else self._resolved_path / mag.path.name
+            if not mag.is_foreign:
+                mag._properties.path = str(
+                    self.resolved_path.relative_to(self.dataset.resolved_path)
+                    / mag.path.name
                 )
+            else:
+                assert mag._properties.path is not None  # for type checking
+            mag._path = (
+                enrich_path(mag._properties.path, self.dataset.resolved_path)
+                if mag._properties.path is not None
+                else self._resolved_path / mag.path.name
+            )
             # Deleting the dataset will close the file handle.
             # The new dataset will be opened automatically when needed.
             del mag._array
@@ -621,7 +581,7 @@ class Layer:
         )
         if shard_shape is None:
             if self.data_format == DataFormat.Zarr:
-                shard_shape = DEFAULT_CHUNK_SHAPE
+                shard_shape = chunk_shape
             else:
                 shard_shape = DEFAULT_SHARD_SHAPE
 
@@ -668,9 +628,7 @@ class Layer:
                     if mag_array_info.data_format in (DataFormat.Zarr, DataFormat.Zarr3)
                     else None
                 ),
-                path=_dump_mag_path(mag_path, self.dataset.resolved_path)
-                if self.dataset.properties.version == 2
-                else None,
+                path=dump_path(mag_path, self.dataset.resolved_path),
             )
         ]
 
@@ -708,13 +666,11 @@ class Layer:
         self._setup_mag(mag, mag_path=mag_path, read_only=read_only)
         mag_view = self._mags[mag]
         mag_array_info = mag_view.info
-        stored_path = None
-        if self.dataset.properties.version == 2:
-            stored_path = (
-                override_stored_path
-                if override_stored_path is not None
-                else _dump_mag_path(mag_path, self.dataset.resolved_path)
-            )
+        stored_path = (
+            override_stored_path
+            if override_stored_path is not None
+            else dump_path(mag_path, self.dataset.resolved_path)
+        )
         self._properties.mags.append(
             MagViewProperties(
                 mag=mag,
@@ -785,7 +741,7 @@ class Layer:
         self._properties.mags.append(
             MagViewProperties(
                 mag=mag_view.mag,
-                path=_dump_mag_path(mag_view.path, self.dataset.resolved_path),
+                path=dump_path(mag_view.path, self.dataset.resolved_path),
                 cube_length=mag_view._properties.cube_length,
                 axis_order=mag_view._properties.axis_order,
             )
@@ -862,6 +818,7 @@ class Layer:
             raise IndexError(
                 f"Deleting mag {mag} failed. There is no mag with this name"
             )
+        mag_view = self.get_mag(mag)
 
         full_path = self._mags[mag].path
         del self._mags[mag]
@@ -869,8 +826,17 @@ class Layer:
             res for res in self._properties.mags if Mag(res.mag) != mag
         ]
         self.dataset._export_as_json()
-        # delete files on disk
-        rmtree(full_path)
+        if not mag_view.is_foreign:
+            # delete files on disk
+            rmtree(full_path)
+        else:
+            # delete symlinks only
+            short_mag_file_path = self.path / mag.to_layer_name()
+            long_mag_file_path = self.path / mag.to_long_layer_name()
+            if short_mag_file_path.exists():
+                short_mag_file_path.unlink()
+            elif long_mag_file_path.exists():
+                long_mag_file_path.unlink()
 
     def add_copy_mag(
         self,
@@ -1639,6 +1605,7 @@ class SegmentationLayer(Layer):
     - Always uses the SEGMENTATION_CATEGORY category type
     - Tracks the largest segment ID present in the data
     - Provides methods for updating the largest segment ID
+    - Adds an `attachments` property for managing attachment files
 
     Attributes:
         largest_segment_id (int | None): Highest segment ID present in data, or None if empty
@@ -1650,6 +1617,16 @@ class SegmentationLayer(Layer):
     """
 
     _properties: SegmentationLayerProperties
+    _attachments: Attachments
+
+    def __init__(
+        self,
+        dataset: "Dataset",
+        properties: SegmentationLayerProperties,
+        read_only: bool,
+    ):
+        super().__init__(dataset, properties, read_only)
+        self._attachments = Attachments(self, properties.attachments)
 
     @property
     def largest_segment_id(self) -> int | None:
@@ -1693,6 +1670,31 @@ class SegmentationLayer(Layer):
     @property
     def category(self) -> LayerCategoryType:
         return SEGMENTATION_CATEGORY
+
+    @property
+    def attachments(self) -> Attachments:
+        """Access, add and remove the attachments of this layer.
+
+        Attachments are additional files that can be attached to a segmentation layer.
+        They can be used to store additional information, such as meshes, agglomerations, segment indices, cumsums and connectomes.
+
+        Examples:
+            ```
+            # Add a mesh attachment to the segmentation layer
+            layer.attachments.add_mesh(
+                mesh_path,
+                name="meshfile",
+                data_format=AttachmentDataFormat.Zarr3,
+            )
+
+            # Access the mesh attachment path
+            layer.attachments.meshes[0].path
+
+            # Remove the mesh attachment
+            layer.attachments.delete_attachment(layer.attachments.meshes[0])
+            ```
+        """
+        return self._attachments
 
     def _get_largest_segment_id_maybe(self) -> int | None:
         return self.largest_segment_id